+++ src/main/java/chess/board/Point.java
@@ -1,4 +1,4 @@
-package chess;
+package chess.board;
 
 import java.util.ArrayList;
 import java.util.HashMap;
