+++ src/main/java/chess/piece/movingstrategy/BishopMovingStrategy.java
@@ -1,4 +1,4 @@
-package chess;
+package chess.piece.movingstrategy;
 
 public class BishopMovingStrategy implements MovingStrategy {
