+++ src/main/java/chess/piece/movingstrategy/EmptyMovingStrategy.java
@@ -1,4 +1,4 @@
-package chess;
+package chess.piece.movingstrategy;
 
 public class EmptyMovingStrategy implements MovingStrategy {
