+++ src/main/java/chess/piece/movingstrategy/KnightMovingStrategy.java
@@ -1,4 +1,4 @@
-package chess;
+package chess.piece.movingstrategy;
 
 public class KnightMovingStrategy implements MovingStrategy {
