+++ src/main/java/chess/piece/movingstrategy/PawnMovingStrategy.java
@@ -1,4 +1,4 @@
-package chess;
+package chess.piece.movingstrategy;
 
 public class PawnMovingStrategy implements MovingStrategy {
