+++ src/main/java/chess/piece/movingstrategy/RookMovingStrategy.java
@@ -1,4 +1,4 @@
-package chess;
+package chess.piece.movingstrategy;
 
 public class RookMovingStrategy implements MovingStrategy {
